+++ src/Kafka/Client.php
@@ -993,7 +993,8 @@ class Client
                     $responsePartition->lastStableOffset,
                     $responsePartition->logStartOffset,
                     $responsePartition->abortedTransactions,
-                    $responsePartition->preferredReadReplica
+                    $responsePartition->preferredReadReplica,
+                    $responsePartition->divergingEpoch
                 );
             }
         }